Tanya| move common methods to drug chart utils
