+++ src/features/DisplayControls/DrugChart/utils/DrugChartUtils.js
@@ -1,11 +1,13 @@
 import axios from "axios";
 import moment from "moment";
+import React from "react";
 import {
   MEDICATIONS_BASE_URL,
   performerFunction,
   asNeededPlaceholderConceptName,
 } from "../../../../constants";
 import _ from "lodash";
+import { FormattedMessage } from "react-intl";
 
 export const fetchMedications = async (
   patientUuid,
@@ -344,3 +346,43 @@ export const getPreviousShiftDetails = (
     previousShiftIndex,
   };
 };
+
+export const isCurrentShift = (
+  shiftConfig,
+  startDateTimeChange,
+  endDateTimeChange
+) => {
+  const shiftDetailsObj = currentShiftHoursArray(new Date(), shiftConfig);
+  const currentShift = shiftDetailsObj.currentShiftHoursArray;
+  let startDateTimeCurrent = getDateTime(new Date(), currentShift[0]);
+  let endDateTimeCurrent = getDateTime(
+    new Date(),
+    currentShift[currentShift.length - 1] + 1
+  );
+
+  if (startDateTimeCurrent > endDateTimeCurrent) {
+    const d = new Date();
+    const currentHour = d.getHours();
+    if (currentHour > 12) {
+      d.setDate(d.getDate() + 1);
+      endDateTimeCurrent = getDateTime(
+        d,
+        currentShift[currentShift.length - 1] + 1
+      );
+    } else {
+      d.setDate(d.getDate() - 1);
+      startDateTimeCurrent = getDateTime(d, currentShift[0]);
+    }
+  }
+  return (
+    startDateTimeCurrent == startDateTimeChange &&
+    endDateTimeCurrent == endDateTimeChange
+  );
+};
+
+export const NotCurrentShiftMessage = (
+  <FormattedMessage
+    id={"NOT_CURRENT_SHIFT_MESSAGE"}
+    defaultMessage={"You're not viewing the current shift"}
+  />
+);
